Again a lot of changes have been done:
- For the reading the data of the config the classe ConfigData is created
in a seperated file called config.py

- For the visualization a bug is fixed. The bug was that whenever
the vis.py goes into a sleep it is not possible to zoom or to create
tiles. !This bus has been fixed by Karol Actun!

- Three methods sim_to_coords, coords_to_sim, check_coords are moved
from the world.py to the header.py. Because those methods are
independent from the world.py. Based on these change some changes
have to been made on the other files.

- The particle counter has been removed from particle.py. The counting
of the particle and given the particle a number based on the counting
is happening now in world.py.

- Some unnecessary variables have been deleted from the matters.